requirements.txt
sentencepiece==0.2.0
accelerate==1.1.0
apex==0.9.10.dev0
basicsr==1.4.2
deepspeed==0.16.3
diffusers==0.31.0
dlib==19.24.6
einops==0.8.1
facexlib==0.3.0
flux==1.3.5
ftfy==6.3.1
horovod==0.28.1
huggingface_hub==0.26.2
imageio==2.36.0
insightface==0.7.3
ipython==8.12.3
ldm==0.1.3
lpips==0.1.4
matplotlib==3.10.0
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
optimum_quanto==0.2.6
pandas==2.2.3
Pillow==11.1.0
pytorch_lightning==2.5.0.post0
regex==2024.9.11
safetensors==0.5.2
scipy==1.15.2
timm==1.0.14
torch==2.5.1
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.20.1
tqdm==4.66.2
transformers==4.46.1
xformers==0.0.28.post3
gradio==5.20.1
